+++ gameboy-wasm/src/canvas.rs
@@ -1,5 +1,4 @@
-use std::sync::Arc;
-
+use std::{rc::Rc, cell::RefCell};
 use yew::prelude::*;
 use gloo::timers::callback::Interval;
 use web_sys::{
@@ -7,68 +6,26 @@ use web_sys::{
     ImageData,
 };
 use wasm_bindgen::JsCast;
-use gameboy_core::{
-    cpu::CPU,
-    cartridge::open_cartridge,
-};
+use gameboy_core::cartridge::open_cartridge;
+use super::emulator::Emulator;
 
-pub struct Emulator {
-    cpu: CPU,
+pub struct Canvas {
+    emulator: Rc<RefCell<Emulator>>,
     rom_name: String,
     canvas: NodeRef,
-}
-
-impl Emulator {
-
-    fn tick(&mut self) {
-        let mut frame_cycles = 0;
-        while frame_cycles < 69_905 {
-            let cycles = self.cpu.tick();
-            self.cpu.mem.update(cycles);
-            frame_cycles += cycles;
-        }
-    }
-
-    fn is_display_updated(&mut self) -> bool {
-        self.cpu.mem.gpu.check_updated()
-    }
-
-    fn pixel_buffer(&self) -> Vec<u8> {
-        let row_pix = 160 * 4 * 4;
-        let mut buf = vec![0; row_pix * 144 * 4 * 4];
-        for (i, raw) in self.cpu.mem.gpu.pixels.iter().enumerate() {
-
-            let col = i % 160;
-            let row = i / 160;
-            let mut rgba = (raw << 8).to_be_bytes();
-            rgba[3] = 0xFF; // Opacity.
-
-            for (j, c) in rgba.iter().enumerate() {
-                for n in 0..4 {
-                    for m in 0..4 {
-                        buf[
-                            ((col * 4 * 4) + (4 * n)) +     // x
-                            (((row * 4) + m ) * row_pix)    // y
-                            + j                             // offset
-                        ] = *c;
-                    }}}}   
-        buf
-    }
 }
 
 enum Msg {
     Start,
 }
 
-impl Component for Emulator {
+impl Component for Canvas {
     type Message = Msg;
     type Properties = ();
 
     fn create(ctx: &yew::Context<Self>) -> Self {
-        let demo = std::fs::read("./pocket.gb").unwrap();
-        let cart = open_cartridge(demo, None);
         Self {
-            cpu: CPU::new(cart, None),
+            emulator: Rc::new(RefCell::new(Emulator::new())),
             rom_name: "demo".to_string(),
             canvas: NodeRef::default(),
         }
@@ -77,7 +34,9 @@ impl Component for Emulator {
     fn update(&mut self, ctx: &yew::Context<Self>, msg: Self::Message) -> bool {
         match msg {
             Msg::Start => {
+                
                 let node_ref = self.canvas.clone();
+                let mut emu = self.emulator.borrow_mut();
                 let canvas = node_ref.cast::<HtmlCanvasElement>().unwrap();
                 let ctx = canvas
                     .get_context("2d")
@@ -85,13 +44,13 @@ impl Component for Emulator {
                     .unwrap()
                     .dyn_into::<web_sys::CanvasRenderingContext2d>()
                     .unwrap();
-                let interval = Interval::new(16, move || {
 
+                let interval = Interval::new(16, move || {
                     // Tick cpu.
-                    self.tick();
+                    emu.tick();
                     // Update display if neccessary.
-                    if self.is_display_updated() {
-                        let buf = self.pixel_buffer();
+                    if emu.is_display_updated() {
+                        let buf = emu.pixel_buffer();
                         let img_data = ImageData::new_with_u8_clamped_array(
                             wasm_bindgen::Clamped(&buf),
                             160 * 4,    // Wdith.
